Moved ResponseHandler and ClientThread into a new File and added the file_manager file
