+++ .gitignore
@@ -1,4 +1,3 @@
-client.py
 marvin/
 docs/
 .vs/
